+++ src/main/java/com/example/logistic/repository/TenantRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TenantRepository extends JpaRepository<Tenant, Integer> {
+public interface TenantRepository extends JpaRepository<Tenant, Long> {
 
 }
